Move oidc rest package to match its new URI
